🐞[fix] WS설정-WS 프로필 관리 페이지 변경 #165
- 탈퇴, 초기화 버튼이 WS 설정 모달에만 존재하도록 수정
